[8028] execd as wazuh module (#8245)
* Initial changes from execd to wmodules movement
Configuration updates and creation of new needed execd files.

* mend

* Updated logs with new wmodule TAG.
Updated all installers appropriately.
Removed unnecessary code and added new needed one.
Makefile update.
Update include file names.

* Fixed UT include

* Fixing server compilation includes

* Start forced exec module in server installation

* Configuration read changes, and manager force execd execution

* Change queue host process from execd to agentd

* Clean up unused code. C Coding Style applied. Doxygen comments added.

* Add shutdown process for new execd as wazuh module

* Fixed execd_shutdown function exposure in Windows.

* Add agent wmodules health status

* Add health status support for server environment.

* Add infinite loop for windows execd module

* Refactor receiver packet process

* Fix windows execd issues

* Fixed wmodule logs for execd. Fixed Makefile command. Added wrap functions for UTs purposes.

* Fix UTs based on execd to wazuh-modules work. Still in progress.

* Fixed Agent/Manager UTs in non Windows platforms.

* Fixed windows UTs

* Fixed compilation issues

* Fix Windows UTs

* PR rework

+++ src/init/update.sh
@@ -408,8 +408,16 @@ UpdateStopOSSEC()
     if [ -d "$PREINSTALLEDDIR/bin/wazuh-syscheckd" ]; then
         rm -rf $PREINSTALLEDDIR/bin/wazuh-syscheckd > /dev/null 2>&1
     fi
-    # Deleting wazuh-logcollector from previous versions from 5.0
-    rm -rf $PREINSTALLEDDIR/bin/wazuh-logcollector > /dev/null 2>&1
+
+    # Deleting wazuh-logcollector if exists (it was migrated to Wazuh modules in v5.0)
+    if [ -d "$PREINSTALLEDDIR/bin/wazuh-logcollector" ]; then
+        rm -rf $PREINSTALLEDDIR/bin/wazuh-logcollector > /dev/null 2>&1
+    fi
+
+    # Deleting wazuh-execd if exists (it was migrated to Wazuh modules in v5.0)
+    if [ -d "$PREINSTALLEDDIR/bin/wazuh-execd" ]; then
+        rm -rf $PREINSTALLEDDIR/bin/wazuh-execd > /dev/null 2>&1
+    fi
 }
 
 UpdateOldVersions()
